requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-sse==2.2.0
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
evdev==1.9.1
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
multidict==6.1.0
numpy==2.2.3
onnxruntime==1.20.1
packaging==24.2
propcache==0.3.0
protobuf==5.29.3
pycparser==2.22
pynput==1.7.8
pyperclip==1.9.0
python-xlib==0.33
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
